[CORRECT] redo role authorization process

+++ OG-ActiTrack_server/src/main/java/com/andres_k/og/controllers/admin/UserController.java
@@ -1,50 +1,53 @@
 package com.andres_k.og.controllers.admin;
 
 import com.andres_k.og.config.HttpResponse;
-import com.andres_k.og.config.Restricted;
 import com.andres_k.og.models.auth.ERoles;
 import com.andres_k.og.models.auth.User;
+import com.andres_k.og.services.AuthorizationService;
 import com.andres_k.og.services.UserService;
-import com.andres_k.og.utils.tools.TJson;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
-@Restricted(tokenRequired = false, roles = ERoles.ADMIN)
 @Controller("AdminUserController")
 @RequestMapping(value = "/admin")
 public class UserController {
     @Autowired
     private UserService userService;
+    @Autowired
+    private AuthorizationService authorizationService;
 
     @RequestMapping(value = "/user/delete", method = RequestMethod.DELETE)
     @ResponseBody
-    public String deleteUser(@RequestBody Long userId) {
+    public ResponseEntity<?> deleteUser(@RequestHeader String Authorization, @RequestBody Long userId) {
         HttpResponse response = new HttpResponse();
 
         try {
+            if (!this.authorizationService.isAuthorized(ERoles.ADMIN, Authorization))
+                return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
             this.userService.deleteUser(userId);
             response.addResult(true);
         } catch (Exception ex) {
             response.addError("Error deleting the user:" + ex.toString());
         }
-        return TJson.toString(response);
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/user/update", method = RequestMethod.PUT)
     @ResponseBody
-    public String update(@RequestBody User user) {
+    public ResponseEntity<?> update(@RequestHeader String Authorization, @RequestBody User user) {
         HttpResponse response = new HttpResponse();
 
         try {
+            if (!this.authorizationService.isAuthorized(ERoles.ADMIN, Authorization))
+                return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
             User newUser = this.userService.updateUser(user);
             response.addResult(newUser);
         } catch (Exception ex) {
             response.addError("Error updating the user:" + ex.toString());
         }
-        return TJson.toString(response);
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
